fix(config): change order of searching context in find_context method
The order of searching context in the find_context method has been changed.
The primary context (no context ≒ WindowManager or OS level) is now searched first, followed by a complete match of config[:context] and request_context, and then a partial match of config[:context] and request_context.
The previous order was a complete match, partial match, and then no context.

With this change, we now treat the primary context as the place to describe the default behavior of like a Window Manager.
Therefore, it is no longer possible to **override** these behaviors in a specific context, such as context: { application: Something }.
This is the intended behavior and should avoid the appearance that the same gesture has multiple behaviors. This can be confusing to the user.

BREAKING CHANGE: no longer possible to **override** primary context with specific context