docs: Add manual method for autostarting fusuma
Updated the README to include instructions for manually creating a desktop entry
to autostart fusuma. This method can be used as an alternative to the
gnome-session-properties approach, addressing compatibility issues on systems
where gnome-session-properties is unavailable. This addition provides a clear
guide for setting up fusuma to start automatically when a session is initiated.

close https://github.com/iberianpig/fusuma/issues/314