fix: handle IOError during process shutdown
- rescue IOError in Process.kill loop
- ignore closed streams on Ctrl-C interrupt

The following error was observed when running Fusuma and pressing Ctrl+C:

```
^C#<Thread:0x00007ed46e788c98 /home/iberianpig/.rbenv/versions/3.3.6/lib/ruby/3.3.0/open3.rb:664 run> terminated with exception (report_on_exception is true):
/home/iberianpig/.rbenv/versions/3.3.6/lib/ruby/3.3.0/open3.rb:664:in `read': stream closed in another thread (IOError)
        from /home/iberianpig/.rbenv/versions/3.3.6/lib/ruby/3.3.0/open3.rb:664:in `block (2 levels) in capture3'
/home/iberianpig/.rbenv/versions/3.3.6/lib/ruby/3.3.0/open3.rb:665:in `read': stream closed in another thread (IOError)
        from /home/iberianpig/.rbenv/versions/3.3.6/lib/ruby/3.3.0/open3.rb:665:in `block (2 levels) in capture3'
```

This error occurs when Fusuma is using `Open3.capture3` to run a command,
and the command is interrupted (e.g., by pressing Ctrl+C).
The internal threads that read from the command's standard output and error
streams attempt to read from a closed stream, resulting in an `IOError`.